Fix 'function-arglist' so it returns only the unique arg list names.
I ran into a small problem with a method that declared defaults for a
some &key values. Swank seems to return these twice if there's a
generic function declared for the method. I.e. I had something like
this:

   (defgeneric set-server (client &key host port path)
    ..)

and:

   (defmethod set-server ((client sphinx-client) &key (host "localhost") (port 3312) path)
     .. )

Which swank returned the arg-list for as:

   (CLIENT &KEY
    HOST
    PORT
    PATH
    (HOST "localhost")
    (PORT 3312))

Which gave me the following lambda list when run through atdoc:

    set-server (client &key host port path (host "localhost") (port 3312))

This fix changes 'function-arglist' so that it returns
a unique list of args, without the defaults. This gives the following
for the above method:

    set-server (client &key host port path)